Update business value class to match backend type
The `is_gql_type` function compares the SDK class name to the
`__typename` received from the backend. To make that work the class
names between SDK and backend have to match. That wasn't the case for
the business value configuration, causing business value to be
unintialized on model creation.

This commit updates the class name to match the backend. Allowing
business value to be configured correctly.